Rewrite HashMap to facilitate growth, alloc macros
HashMap is now a value type so that only the data pointer needs
to be changed when the map is grown. This avoids awkward call-site
syntax where the map pointer had to be reassigned with each insert
call.

There are now macros to ease allocating, reallocating, deallocating
memory, arrays and flexible array member structs.