Fix locals corruption error
When starting an orbit invocation, the stack pointer was set at
the start of the task's stack, even if the entry point function
defined any variables. This caused variables to be overwritten by
temporary stack values in some case. The stack pointer is now set
after the entry point's locals table in orbit_gcTaskNew().